Fix flow of purchase

+++ src/Services/StockMarketService/Commands.cs
@@ -21,13 +21,14 @@ namespace StockMarketService
             this.context = context;
             this.manager = manager;
         }
-        public void AddNewStock(Stock stock) {
+        public void AddNewStock(Stock stock)
+        {
             Console.WriteLine(JsonSerializer.Serialize(stock));
             context.Stocks.Add(stock);
             context.SaveChanges();
-            
+
             UpdateClients();
-            
+
         }
 
         public void UpdateClients()
@@ -36,47 +37,55 @@ namespace StockMarketService
             manager.UpdateAllClients(allStocks);
         }
 
-        public void UpdateStockPrice(string stockName,decimal newStockPrice) {
+        public void UpdateStockPrice(string stockName, int newStockPrice)
+        {
             var stock = context.Stocks.Find(stockName);
             stock.HistoricPrice.Add(new StockPrice(newStockPrice, DateTime.Now));
             context.Stocks.Update(stock);
             UpdateClients();
         }
 
-        public List<Stock> GetStocks() {
+        public List<Stock> GetStocks()
+        {
             return context.Stocks.Include(s => s.HistoricPrice).Include(s => s.Seller).ToList();
         }
-        
-        public Stock GetStock(string name) {
+
+        public Stock GetStock(string name)
+        {
             return context.Stocks.Include(s => s.HistoricPrice).Include(s => s.Seller).FirstOrDefault(s => s.Name == name);
         }
-        
+
         public List<Stock> GetUserStock(string userid)
         {
             return context.Stocks.Where(e => e.Seller.FirstOrDefault(s => s.SellerId == userid) != null).Include(s => s.HistoricPrice).Include(s => s.Seller).ToList();
         }
-        
-        public void UpdateStock(Stock stock) {
+
+        public void UpdateStock(Stock stock)
+        {
             Console.WriteLine(JsonSerializer.Serialize(stock));
             context.Stocks.Update(stock);
             context.SaveChanges();
             UpdateClients();
         }
-        
-        public Seller GetSeller(string name, int number) {
+
+        public Seller GetSeller(string name, int number)
+        {
             Stock stock = GetStock(name);
             Seller seller = stock.Seller.FirstOrDefault(s => s.SellingAmount >= number);
-            if (seller == null ) {
+            if (seller == null)
+            {
                 return null;
             }
             return seller;
         }
 
-        public bool SellStock(string name, int number, string sellerid) {
+        public bool SellStock(string name, int number, string sellerid)
+        {
             Stock stock = GetStock(name);
             context.Attach(stock);
             Seller seller = stock.Seller.FirstOrDefault(s => s.SellerId == sellerid);
-            if (seller == null ) {
+            if (seller == null)
+            {
                 return false;
             }
 
@@ -84,12 +93,14 @@ namespace StockMarketService
             context.SaveChanges();
             return true;
         }
-        
-        public bool BuyStock(string name, int number) {
+
+        public bool BuyStock(string name, int number)
+        {
             Stock stock = GetStock(name);
             context.Attach(stock);
             Seller seller = stock.Seller.FirstOrDefault(s => s.SellingAmount >= number);
-            if (seller == null ) {
+            if (seller == null)
+            {
                 return false;
             }
             seller.SellingAmount -= number;
